CMD-20 Change class name + adjust other methods to fit the goal of this class
